fix(vue): properly handle filtered table nodes
