+++ packages/components-vue/src/types/props/index.ts
@@ -1,0 +1,4 @@
+export * from "./base";
+export * from "./modal";
+export * from "./table";
+export * from "./value";
